Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pillow
transformers
torch
torchvision
numpy
pandas
timm==0.4.12